# gi_service/contract/rest/organisation_api_contract.yaml
openapi: 3.1.0

info:
  title: OpenGIN Service API - Organization
  description: This document gives the structure of APIs for the OpenGIN Organization Services.
  version: 1.0.0

servers:
  - url: http://localhost:8081

tags:
  - name: Organization
    description: APIs for organization
  - name: Data
    description: APIs for data

paths:
  /v1/presidents:
    get:
      tags:
        - Organization
      summary: Get all presidents with their term dates and gazettes sorted by date.
      description: Returns a sorted list of presidents with their term dates and corresponding gazette ids and dates.
      operationId: getPresidents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      $ref: "#/components/schemas/President"
              example:
                body:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03-01_cit_1
                    tenureList:
                      - startDate: 2024-09-23T00:00:00Z
                        endDate: ""
                        gazetteList:
                          - date: 2024-11-18T00:00:00Z
                            idList:
                              - 2411-09
                          - date: 2024-12-13T00:00:00Z
                            idList:
                              - 2412-03
                              - 2412-04
        "500":
          description: Server Error

  /v1/organisation/prime-minister:
    post:
      tags:
        - Organization
      summary: Get prime minister Data
      description: Returns the prime minister details based on the given date.
      operationId: getPrimeMinister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    required:
                      - name
                      - id
                    properties:
                      name:
                        type: string
                      id:
                        type: string
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/organisation/cabinet-flow/{president_id}:
    post:
      tags:
        - Organization
      summary: Get cabinet flow for a president (How the organisation structure changed over time).
      description: >-
        Returns d3-force-directed graph data with nodes, links, and dates.
        Nodes correspond to ministers; links connect minister nodes across
        consecutive included dates, with a value equal to the number of
        departments that moved between them. The dates array reports the
        processing status for every requested date. Only dates with
        status "ok" contribute to nodes and links; dates with status
        "no_data" (departmentsCount 0) or "error" are excluded from the
        chart. When a middle date is excluded, links bridge directly
        between the nearest included dates on either side.
      operationId: getCabinetFlow
      parameters:
        - name: president_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: List of dates (max 10) for which to compute the cabinet flow.
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
            example:
              - "2024-09-23"
              - "2024-09-24"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - name
                        - time
                      properties:
                        id:
                          type: string
                          description: Unique identifier for the ministry node.
                        name:
                          type: string
                          description: Display name of the ministry.
                        time:
                          type: string
                          description: The date snapshot this node belongs to (YYYY-MM-DD).
                  links:
                    type: array
                    items:
                      type: object
                      required:
                        - source
                        - target
                        - value
                        - departmentIds
                      properties:
                        source:
                          type: integer
                          description: Index of the source node in the nodes array.
                        target:
                          type: integer
                          description: Index of the target node in the nodes array.
                        value:
                          type: integer
                          description: Number of departments shared between the source and target ministry nodes.
                        departmentIds:
                          type: array
                          items:
                            type: string
                          description: IDs of the departments that moved along this link between consecutive dates.
                  dates:
                    type: array
                    description: >-
                      One entry per requested date. Entries with status "ok"
                      are included in nodes and links; entries with status
                      "no_data" or "error" are reported here but omitted
                      from the chart.
                    items:
                      type: object
                      required:
                        - date
                        - status
                      properties:
                        date:
                          type: string
                          description: The date snapshot (YYYY-MM-DD).
                        status:
                          type: string
                          description: >-
                            Processing status for this date. "ok" — data
                            retrieved and included in nodes/links.
                            "no_data" — no departments on this date
                            (departmentsCount 0); excluded from nodes/links.
                            "error" — retrieval failed; excluded from
                            nodes/links.
                        departmentsCount:
                          type: integer
                          description: >-
                            Total number of active departments on this date.
                            Present and 0 when status is "no_data"; omitted
                            when status is "error".
              example:
                nodes:
                  - id: "2289-43_min_1"
                    name: "Minister of Defence"
                    time: "2024-01-01"
                  - id: "2289-43_min_1"
                    name: "Minister of Defence"
                    time: "2024-03-01"
                links:
                  - source: 0
                    target: 1
                    value: 25
                    departmentIds:
                      - "2289-43_dep_1"
                      - "2289-43_dep_2"
                dates:
                  - date: "2024-01-01"
                    status: "ok"
                    departmentsCount: 449
                  - date: "2024-02-01"
                    status: "no_data"
                    departmentsCount: 0
                  - date: "2024-03-01"
                    status: "ok"
                    departmentsCount: 449
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/person-history/{personId}:
    get:
      tags:
        - Organization
      summary: get person history
      description: Returns the history of the person.
      operationId: getPersonHistory
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      properties:
                        startDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        endDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        isPresident:
                          type: boolean
                        portfolio:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
              example:
                body:
                  - startDate:
                      date: 2024-11-18
                    endDate:
                      date: Present
                    isPresident: true
                    portfolio:
                      name: Minister of Digital Economy
                      id: 2434-min-1
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/department-history/{departmentId}:
    get:
      tags:
        - Organization
      summary: Get department history.
      description: Returns the history of the department.
      operationId: getDepartmentHistory
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      properties:
                        startDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        endDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        portfolio:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
              example:
                body:
                  - startDate:
                      date: 2024-11-18
                    endDate:
                      date: Present
                    portfolio:
                      name: Minister of Plantation and Community Infrastructure
                      id: 2434-min-2
                      ministers:
                        - name: Samantha Vidyarathna
                          id: 2405-cit-4
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/organisation/entity-names:
    post:
      tags:
        - Organization
      summary: Resolve entity IDs to display names.
      description: Returns a dictionary mapping each entity ID to its decoded display name. IDs that cannot be resolved are omitted from the response.
      operationId: resolveEntityNames
      requestBody:
        required: true
        description: List of entity IDs to resolve.
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
            example:
              - "2289-43_min_1"
              - "2289-43_dep_1"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                description: Map of entity ID to display name.
              example:
                "2289-43_min_1": "Minister of Defence"
                "2289-43_dep_1": "Ministry of Health"
        "400":
          description: Bad request
        "500":
          description: Server Error

  /v1/organisation/department-history/{department_id}:
    get:
      tags:
        - Organization
      summary: Get department history timeline.
      description: Returns a timeline of a department including ministry relations and ministers.
      operationId: getDepartmentHistoryTimeline
      parameters:
        - name: department_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ministry_id:
                      type: string
                    ministry_name:
                      type: string
                    minister_id:
                      type: string
                    minister_name:
                      type: string
                    period:
                      type: string
              example:
                - ministry_id: "min_01"
                  ministry_name: "Ministry of Finance"
                  minister_id: "pers_01"
                  minister_name: "John Doe"
                  period: "2020 Jan 01 - 2021 Jan 01"
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/active-portfolio-list:
    post:
      tags:
        - Organization
      summary: Get active portfolio list.
      description: Returns a list of portfolios under a given president and a given date.
      operationId: getActivePortfolioList
      parameters:
        - name: presidentId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrive the list of active portfolios.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    properties:
                      portfolioList:
                        type: array
                        items:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            isNew:
                              type: boolean
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
                                  isPresident:
                                    type: boolean
                                  isNew:
                                    type: boolean
                      activeMinistries:
                        type: integer
                      newMinistries:
                        type: integer
                      newMinisters:
                        type: integer
                      ministriesUnderPresident:
                        type: integer
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/gazettes-by-president/{presidentId}:
    post:
      tags:
        - Organization
      summary: Get a list of gazettes for a given time range and president.
      description: Returns a list of gazettes for a given time range and president, sort the values according to the date.
      operationId: getGazettesByPresident
      parameters:
        - name: presidentId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date range used to retrieve the gazette list for the given president term and date range.
        content:
          application/json:
            schema:
              type: object
              description: Required both startDate and endDate for the range.
              properties:
                startDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
                endDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      additionalProperties:
                        type: array
                        items:
                          type: object
                          required:
                            - gazetteId
                            - publishedDate
                          properties:
                            gazetteId:
                              type: string
                            publishedDate:
                              type: object
                              required:
                                - date
                              properties:
                                date:
                                  type: string
              example:
                body:
                  - "2025-04-23":
                      - gazetteId: 2411-09
                        publishedDate:
                          date: 2024-11-23T00:00:00Z
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-23T00:00:00Z
                  - "2025-04-24":
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-24T00:00:00Z
                  - "2025-04-25":
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-25T00:00:00Z
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/departments-by-portfolio/{portfolio_id}:
    post:
      tags:
        - Organization
      summary: Get active departments for a portfolio.
      description: Returns a list of departments under a given portfolio and a given date.
      operationId: getDepartmentsByPortfolio
      parameters:
        - name: portfolio_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrive the list of active departments.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - id
                      properties:
                        name:
                          type: string
                        id:
                          type: string
                        isNew:
                          type: boolean
                        hasData:
                          type: boolean

              example:
                body:
                  - totalDepartments: 2
                    newDepartments: 2
                    departmentList:
                      - name: National Police Academy
                        id: 2187-27_dep_4
                        isNew: false
                        hasData: true
                      - name: Assets vested to the Secretary to the Treasury, under the Act, No. 43 of 2011
                        id: 2153-12_dep_437
                        isNew: false
                        hasData: false
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/persons-by-portfolio/{portfolioId}:
    post:
      tags:
        - Organization
      summary: Get active persons for a portfolio.
      description: Returns a list of persons under a given portfolio and a given date.
      operationId: getPersonsByMinistry
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrieve the list of active persons.
        content:
          application/json:
            schema:
              type: object
              description: required date
              properties:
                activeDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - id
                      properties:
                        name:
                          type: string
                        id:
                          type: string
                        isPresident:
                          type: boolean
                        isNew:
                          type: boolean

              example:
                body:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03_cit_1
                    isPresident: true
                    isNew: false
                  - name: Gotabaya Rajapaksa
                    id: 2149-34_cit_1
                    isPresident: false
                    isNew: false

        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

components:
  schemas:
    Date:
      type: object
      required:
        - date
      properties:
        date:
          type: string

    President:
      allOf:
        - type: object
          allOf:
            - $ref: "#/components/schemas/Person"
        - type: object
          required:
            - tenureList
          properties:
            tenureList:
              type: array
              items:
                $ref: "#/components/schemas/Tenure"

    Tenure:
      type: object
      required:
        - startDate
        - endDate
        - gazetteList
      properties:
        startDate:
          type: string
        endDate:
          type: string
        gazetteList:
          type: array
          items:
            type: object
            required:
              - date
              - idList
            properties:
              date:
                type: string
              idList:
                type: array
                items:
                  type: string

    Person:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string

    Portfolio:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string
        ministers:
          type: array
          items:
            type: object
            required:
              - name
              - id
            properties:
              name:
                type: string
              id:
                type: string

    Department:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string